+++ apps/contract/contracts/nft/test-snapshots/test/test-basic-flow.1.json
