+++ apps/contract/contracts/nft/test-snapshots/test/test-error-handling.1.json
