+++ apps/indexer/src/test/mapping-handlers.test.ts
