+++ apps/kyc-server/src/utils/build-client.ts
