+++ apps/web/components/pages/team-members/components/team-member-list.tsx
@@ -7,7 +7,7 @@ import {
 	CardTitle,
 } from '~/components/base/card'
 import type { formInputs } from '../team-members'
-import TeamMemberCard from './TeamMemberCard'
+import TeamMemberCard from './team-member-card'
 
 interface TeamMemberListProps {
 	form: UseFormReturn<formInputs>
