+++ apps/web/components/sections/projects/business-model.tsx
