+++ apps/web/components/sections/projects/competitive-advantages.tsx
