+++ apps/web/components/sections/projects/investment-details.tsx
